+++ packages/webview_flutter/webview_flutter_android/example/lib/web_view.dart
@@ -237,11 +237,10 @@ class WebView extends StatefulWidget {
   /// The default policy is [AutoMediaPlaybackPolicy.require_user_action_for_all_media_types].
   final AutoMediaPlaybackPolicy initialMediaPlaybackPolicy;
 
-  /// The background color of the webview.
+  /// The background color of the [WebView].
   ///
-  /// When null the platform's webview default background color is used.
-  ///
-  /// By default `backgroundColor` is null.
+  /// When `null` the platform's webview default background color is used. By
+  /// default [backgroundColor] is `null`.
   final Color? backgroundColor;
 
   @override
@@ -381,6 +380,14 @@ class WebViewController {
   /// Throws an ArgumentError if the [absoluteFilePath] does not exist.
   Future<void> loadFile(String absoluteFilePath) {
     return _webViewPlatformController.loadFile(absoluteFilePath);
+  }
+
+  /// Loads the Flutter asset specified in the pubspec.yaml file.
+  ///
+  /// Throws an ArgumentError if [key] is not part of the specified assets
+  /// in the pubspec.yaml file.
+  Future<void> loadFlutterAsset(String key) {
+    return _webViewPlatformController.loadFlutterAsset(key);
   }
 
   /// Loads the supplied HTML string.