+++ packages/webview_flutter/webview_flutter_wkwebview/example/lib/main.dart
@@ -91,7 +91,7 @@ class _WebViewExampleState extends State<_WebViewExample> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Colors.green,
+      backgroundColor: const Color(0xFF4CAF50),
       appBar: AppBar(
         title: const Text('Flutter WebView example'),
         // This drop down menu demonstrates that Flutter widgets can be shown over the web view.
@@ -381,9 +381,7 @@ class _SampleMenu extends StatelessWidget {
 
   Future<void> _onTransparentBackground(
       WebViewController controller, BuildContext context) async {
-    final String contentBase64 =
-        base64Encode(const Utf8Encoder().convert(kTransparentBackgroundPage));
-    await controller.loadUrl('data:text/html;base64,$contentBase64');
+    await controller.loadHtmlString(kTransparentBackgroundPage);
   }
 
   static Future<String> _prepareLocalFile() async {